Remove cacheableService / Add cacheable option
