+++ src/Itkg/Consumer/Service/Service.php
@@ -6,6 +6,7 @@ use Itkg\Consumer\Client\ClientInterface;
 use Itkg\Consumer\Event\ConfigEvent;
 use Itkg\Consumer\Event\ServiceEvent;
 use Itkg\Consumer\Event\ServiceEvents;
+use Itkg\Core\CacheableInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Itkg\Consumer\Response;
@@ -18,8 +19,12 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  *
  * @package Itkg\Consumer\Service
  */
-class Service implements ServiceInterface
+class Service implements ServiceInterface, CacheableInterface
 {
+    /**
+     * @var bool
+     */
+    private $loaded = false;
     /**
      * @var Request
      */
@@ -221,6 +226,80 @@ class Service implements ServiceInterface
     }
 
     /**
+     * Hash key getter
+     *
+     * @return string
+     */
+    public function getHashKey()
+    {
+        return strtr($this->getIdentifier(), ' ','_').md5(
+            sprintf(
+                '%s_%s_%s',
+                $this->request->getContent(),
+                $this->request->getUri(),
+                json_encode($this->request->headers->all())
+            )
+        );
+    }
+
+    /**
+     * Get cache TTL
+     *
+     * @return int
+     */
+    public function getTtl()
+    {
+        return $this->options['cache_ttl'];
+    }
+
+    /**
+     * Return if object is already loaded from cache
+     *
+     * @return bool
+     */
+    public function isLoaded()
+    {
+        return $this->loaded;
+    }
+
+    /**
+     * Set is loaded
+     *
+     * @param bool $isLoaded
+     */
+    public function setIsLoaded($isLoaded = true)
+    {
+        $this->loaded = $isLoaded;
+    }
+
+    /**
+     * Get data from service for cache set
+     *
+     * @return mixed
+     */
+    public function getDataForCache()
+    {
+        return call_user_func(
+            $this->options['cache_serializer'],
+            $this->response
+        );
+    }
+
+    /**
+     * Restore data after cache load
+     *
+     * @param $data
+     * @return $this
+     */
+    public function setDataFromCache($data)
+    {
+        $this->response = call_user_func(
+            $this->options['cache_unserializer'],
+            $data
+        );
+    }
+
+    /**
      * Manage configuration
      *
      * @param array $options
@@ -240,9 +319,13 @@ class Service implements ServiceInterface
                 'logger'
             ))
             ->setDefaults(array(
-                'response_format' => 'json', // Define a format used by serializer (json, xml, etc),
-                'response_type'   => 'array', // Define a mapped class for response content deserialization,
-                'loggable'        => false
+                'response_format'    => 'json', // Define a format used by serializer (json, xml, etc),
+                'response_type'      => 'array', // Define a mapped class for response content deserialization,
+                'loggable'           => false,
+                'cacheable'          => false,
+                'cache_ttl'          => null,
+                'cache_serializer'   => 'serialize',
+                'cache_unserializer' => 'unserialize'
             )
         );
